refactor: extract converters/base/helpers/ subpackage (#310)
Reorganize converters/base/ to separate abstract base classes (the Ops
pattern contract) from utility functions (implementation details):

- Move cache.py, schema.py, tool_content.py, image_limit.py into helpers/
- Extract orphan fix utilities from tools.py into helpers/orphan_fix.py
- Move tool_call_unwind.py from shims/providers/argo/utils/ to helpers/
- tools.py now contains only the BaseToolOps abstract class
- helpers/__init__.py re-exports public functions for clean imports
- Update all import sites across 4 converters, gateway, and tests
- Update README file structure diagrams (en + zh)
