+++ src/llm_rosetta/converters/base/helpers/cache.py
