+++ src/llm_rosetta/converters/base/helpers/image_limit.py
