+++ src/llm_rosetta/converters/base/helpers/schema.py
