+++ src/llm_rosetta/converters/base/helpers/tool_content.py
