+++ src/llm_rosetta/converters/openai_chat/converter.py
@@ -35,7 +35,7 @@ from ...types.ir.stream import (
 )
 from ..base import BaseConverter
 from ..base.context import ConversionContext, StreamContext
-from ..base.tools import fix_orphaned_tool_calls_ir, strip_orphaned_tool_config
+from ..base.helpers import fix_orphaned_tool_calls_ir, strip_orphaned_tool_config
 from ._constants import OPENAI_CHAT_REASON_FROM_PROVIDER, OPENAI_CHAT_REASON_TO_PROVIDER
 from .config_ops import OpenAIChatConfigOps
 from .content_ops import OpenAIChatContentOps
