--- src/llm_rosetta/shims/providers/argo/utils/__init__.py
@@ -1,6 +1,0 @@
-"""Argo-specific utility modules.
-
-Shared helpers for the Argo provider shims (``argo--openai_chat``,
-``argo--anthropic``).  These utilities address Argo gateway quirks that
-are not part of the base conversion logic.
-"""
