+++ tests/converters/base/test_tool_content.py
@@ -2,7 +2,7 @@
 Tests for the shared tool content conversion helpers.
 """
 
-from llm_rosetta.converters.base.tool_content import (
+from llm_rosetta.converters.base.helpers.tool_content import (
     convert_content_blocks_to_ir,
     convert_ir_content_blocks_to_p,
 )
