+++ tests/converters/base/test_tools_sanitize.py
@@ -1,6 +1,9 @@
 """Tests for sanitize_schema and _flatten_combination in converters/base/schema.py."""
 
-from llm_rosetta.converters.base.schema import sanitize_schema, _flatten_combination
+from llm_rosetta.converters.base.helpers.schema import (
+    sanitize_schema,
+    _flatten_combination,
+)
 
 
 class TestFlattenCombination:
