+++ tests/test_tool_call_unwind.py
@@ -2,7 +2,7 @@
 
 from typing import Any
 
-from llm_rosetta.shims.providers.argo.utils.tool_call_unwind import (
+from llm_rosetta.converters.base.helpers.tool_call_unwind import (
     unwind_parallel_tool_calls_ir,
 )
